+++ packages/react-hooks/src/useOwnEraRewards.ts
@@ -3,6 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { DeriveStakerReward } from '@polkadot/api-derive/types';
+import { EraIndex } from '@polkadot/types/interfaces';
 
 import { useEffect, useState } from 'react';
 
@@ -13,6 +14,7 @@ import { useOwnStashIds } from './useOwnStashes';
 
 interface OwnRewards {
   allRewards?: Record<string, DeriveStakerReward[]>;
+  isLoadingRewards: boolean;
   rewardCount: number;
 }
 
@@ -25,22 +27,35 @@ function getRewards ([[stashIds], available]: [[string[]], DeriveStakerReward[][
 
   return {
     allRewards,
+    isLoadingRewards: false,
     rewardCount: Object.values(allRewards).filter((rewards) => rewards.length !== 0).length
   };
 }
 
-export default function useOwnEraRewards (): OwnRewards {
+export default function useOwnEraRewards (maxEras = 1000): OwnRewards {
   const { api } = useApi();
   const mountedRef = useIsMountedRef();
   const stashIds = useOwnStashIds();
-  const available = useCall<[[string[]], DeriveStakerReward[][]]>(stashIds && api.derive.staking?.stakerRewardsMulti, [stashIds], { withParams: true });
-  const [state, setState] = useState<OwnRewards>({ rewardCount: 0 });
+  const allEras = useCall<EraIndex[]>(api.derive.staking?.erasHistoric, []);
+  const [filteredEras, setFilteredEras] = useState<EraIndex[]>([]);
+  const available = useCall<[[string[]], DeriveStakerReward[][]]>((filteredEras?.length > 0) && stashIds && api.derive.staking?.stakerRewardsMultiEras, [stashIds, filteredEras], { withParams: true });
+  const [state, setState] = useState<OwnRewards>({ isLoadingRewards: true, rewardCount: 0 });
+
+  useEffect((): void => {
+    setState({ isLoadingRewards: true, rewardCount: 0 });
+  }, [maxEras]);
 
   useEffect((): void => {
     mountedRef.current && available && setState(
       getRewards(available)
     );
   }, [available, mountedRef]);
+
+  useEffect((): void => {
+    allEras && setFilteredEras(
+      allEras.reverse().filter((_, index) => index < maxEras).reverse()
+    );
+  }, [allEras, maxEras]);
 
   return state;
 }